feat: implement registration form and hero components; refactor register page
