--- conexia_front/src/service/users/usersFetch.js
@@ -1,18 +1,0 @@
-import { config } from '../../config';
-
-export const fetchPing = async() => {
-  const response = await fetch(`${config.API_URL}/users/ping`, {
-    cache: 'no-store', 
-    headers: {
-      'Content-Type': 'application/json',
-    },
-  });
-  
-  if (!response.ok) {
-    throw new Error(`HTTP error! status: ${response.status}`);
-  }
-
-  const { data } = (await response.json());
-  return data;
-};
-
